Refactor PollPage to use PollClientPage and update vote count in db.json; add qrcode.react dependency
