+++ db.json
@@ -50,7 +50,7 @@
         {
           "id": 3,
           "text": "JS",
-          "votes": 1
+          "votes": 2
         }
       ],
       "createdBy": 1,
